refactor: move files around
put the oo wrapper modules in a hierarchy
