+++ tests/fixtures/vcr_cassettes/test_fn_decorator/test_fn_decorator.yaml
@@ -1,0 +1,148 @@
+interactions:
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      connection:
+      - keep-alive
+    method: GET
+    uri: http://localhost:8000/v1/functions/by_path/translate
+  response:
+    body:
+      string: '{"path":"translate","instructions":"Operation: translate\n\nOperation
+        description: Translate text to a target language.","id":82,"description":"Translate
+        text to a target language.","input_schema":null,"out_schema":{"type":"string","$defs":{}},"few_shot":false,"few_shot_count":2,"dataset_uuid":"0331a79c-e4f6-44c3-8176-56d908caaf43","use_semantic_search":false,"model":"azure/gpt4-eu","language_model_id":8,"index_ids":[],"linked_function_ids":null,"revision":1,"cache_configuration":{"exact_match_cache_ttl":0,"semantic_cache_threshold":0.95,"semantic_cache_ttl":0}}'
+    headers:
+      content-length:
+      - '568'
+      content-type:
+      - application/json
+      date:
+      - Tue, 07 May 2024 20:11:36 GMT
+      server:
+      - uvicorn
+    status:
+      code: 200
+      message: OK
+- request:
+    body: '{"id": 82, "path": "translate", "description": "Translate text to a target
+      language.", "input_schema": null, "out_schema": {"type": "string", "$defs":
+      {}}, "instructions": "Operation: translate\n\nOperation description: Translate
+      text to a target language.", "model": null, "index_ids": [], "use_semantic_search":
+      null, "few_shot": false, "few_shot_count": null, "cache_configuration": null}'
+    headers:
+      accept:
+      - '*/*'
+      connection:
+      - keep-alive
+      content-length:
+      - '391'
+      content-type:
+      - application/json
+    method: PATCH
+    uri: http://localhost:8000/v1/functions/82
+  response:
+    body:
+      string: '{"path":"translate","instructions":"Operation: translate\n\nOperation
+        description: Translate text to a target language.","id":82,"description":"Translate
+        text to a target language.","input_schema":null,"out_schema":{"type":"string","$defs":{}},"few_shot":false,"few_shot_count":2,"dataset_uuid":"0331a79c-e4f6-44c3-8176-56d908caaf43","use_semantic_search":false,"model":"azure/gpt4-eu","language_model_id":8,"index_ids":[],"linked_function_ids":null,"revision":1,"cache_configuration":{"exact_match_cache_ttl":0,"semantic_cache_threshold":0.95,"semantic_cache_ttl":0}}'
+    headers:
+      content-length:
+      - '568'
+      content-type:
+      - application/json
+      date:
+      - Tue, 07 May 2024 20:11:36 GMT
+      server:
+      - uvicorn
+    status:
+      code: 200
+      message: OK
+- request:
+    body: '{"parent_span_uuid": null, "messages": [{"role": "user", "content": "{\"text\":
+      \"Hello\", \"target_language\": \"es\"}"}]}'
+    headers:
+      accept:
+      - '*/*'
+      connection:
+      - keep-alive
+      content-length:
+      - '123'
+      content-type:
+      - application/json
+    method: POST
+    uri: http://localhost:8000/v1/chat/translate
+  response:
+    body:
+      string: '{"span_id":"0c282ba5-4c82-4b2e-ba84-410535df190f","message":null,"json_payload":"Hola","context":null,"cached":false}'
+    headers:
+      content-length:
+      - '117'
+      content-type:
+      - application/json
+      date:
+      - Tue, 07 May 2024 20:11:36 GMT
+      server:
+      - uvicorn
+    status:
+      code: 200
+      message: OK
+- request:
+    body: '{"parent_span_uuid": null, "messages": [{"role": "user", "content": "{\"text\":
+      \"Hello\", \"target_language\": \"es\"}"}]}'
+    headers:
+      accept:
+      - '*/*'
+      connection:
+      - keep-alive
+      content-length:
+      - '123'
+      content-type:
+      - application/json
+    method: POST
+    uri: http://localhost:8000/v1/chat/translate
+  response:
+    body:
+      string: '{"span_id":"c48b7221-4efc-4ba7-8f6c-fadfda9e996e","message":null,"json_payload":"Hola","context":null,"cached":false}'
+    headers:
+      content-length:
+      - '117'
+      content-type:
+      - application/json
+      date:
+      - Tue, 07 May 2024 20:11:37 GMT
+      server:
+      - uvicorn
+    status:
+      code: 200
+      message: OK
+- request:
+    body: '{"dimension": "metric", "value": 1.0, "comment": null}'
+    headers:
+      accept:
+      - '*/*'
+      connection:
+      - keep-alive
+      content-length:
+      - '54'
+      content-type:
+      - application/json
+    method: POST
+    uri: http://localhost:8000/v1/spans/c48b7221-4efc-4ba7-8f6c-fadfda9e996e/metrics
+  response:
+    body:
+      string: '{"id":"651e4654-0376-4163-8590-ef4d52a2c17e","span_id":"c48b7221-4efc-4ba7-8f6c-fadfda9e996e","dimension":"metric","value":1.0,"comment":null,"updated_at":null,"created_at":"2024-05-07T20:11:38.831035Z"}'
+    headers:
+      content-length:
+      - '203'
+      content-type:
+      - application/json
+      date:
+      - Tue, 07 May 2024 20:11:38 GMT
+      server:
+      - uvicorn
+    status:
+      code: 200
+      message: OK
+version: 1
